Refactor TypeDecl and Invokable classes
