+++ compiler/BIRReader.cpp
@@ -8,6 +8,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 using namespace std;
+using namespace nballerina;
 
 // Instruction readers object declarations
 ReadCondBrInsn ReadCondBrInsn::readCondBrInsn;
