+++ compiler/llvmTypeDeclGen.cpp
@@ -1,14 +1,17 @@
-#include "BIR.h"
 #include "BalTypeDecl.h"
 
-TypeDecl::TypeDecl(int tagp, string namep, int flagsp)
+namespace nballerina {
+// TypeDecl class
+TypeDecl::TypeDecl(int tagp, std::string namep, int flagsp)
     : tag(tagp), name(std::move(namep)), flags(flagsp) {}
 
 int TypeDecl::getTypeTag() { return tag; }
 std::string TypeDecl::getTypeDeclName() { return name; }
 int TypeDecl::getFlags() { return flags; }
 
+// MapTypeDecl class
 MapTypeDecl::MapTypeDecl(int tag, std::string name, int flags, int memberTag)
     : TypeDecl{tag, name, flags}, memberTypeTag{memberTag} {}
 
 int MapTypeDecl::getTypeMemberTag() { return memberTypeTag; }
+} // namespace nballerina
